fix core tests, move wallet dependent tests into integration tests
