+++ chain/Cargo.toml
@@ -20,5 +20,6 @@ grin_store = { path = "../store" }
 grin_util = { path = "../util" }
 
 [dev-dependencies]
+grin_wallet = { path = "../wallet" }
 env_logger = "0.3"
 rand = "0.3"
