+++ core/Cargo.toml
@@ -20,3 +20,6 @@ time = "0.1"
 
 grin_keychain = { path = "../keychain" }
 grin_util = { path = "../util" }
+
+[dev-dependencies]
+grin_wallet = { path = "../wallet" }
