+++ core/src/core/block.rs
@@ -18,8 +18,8 @@ use time;
 use rand::{thread_rng, Rng};
 use std::collections::HashSet;
 
-use core::{Commitment, Committed, Input, KernelFeatures, Output, OutputFeatures, Proof,
-           ProofMessageElements, ShortId, Transaction, TxKernel};
+use core::{Commitment, Committed, Input, KernelFeatures, Output, OutputFeatures, Proof, ShortId,
+           Transaction, TxKernel};
 use consensus;
 use consensus::{exceeds_weight, reward, VerifySortOrder, REWARD};
 use core::hash::{Hash, HashWriter, Hashed, ZERO_HASH};
@@ -30,7 +30,6 @@ use ser::{self, read_and_verify_sorted, Readable, Reader, Writeable, WriteableSo
 use global;
 use keychain;
 use keychain::BlindingFactor;
-use util::kernel_sig_msg;
 use util::LOGGER;
 use util::{secp, secp_static, static_secp_instance};
 
@@ -400,8 +399,6 @@ impl Block {
 	pub fn new(
 		prev: &BlockHeader,
 		txs: Vec<&Transaction>,
-		keychain: &keychain::Keychain,
-		key_id: &keychain::Identifier,
 		difficulty: Difficulty,
 		reward_output: (Output, TxKernel),
 	) -> Result<Block, Error> {
@@ -701,7 +698,8 @@ impl Block {
 		Ok(())
 	}
 
-	fn verify_sums(
+	/// Verify sums
+	pub fn verify_sums(
 		&self,
 		prev_output_sum: &Commitment,
 		prev_kernel_sum: &Commitment,
@@ -768,12 +766,12 @@ impl Block {
 		Ok(sum)
 	}
 
-	// Validate the coinbase outputs generated by miners. Entails 2 main checks:
-	//
-	// * That the sum of all coinbase-marked outputs equal the supply.
-	// * That the sum of blinding factors for all coinbase-marked outputs match
-	//   the coinbase-marked kernels.
-	fn verify_coinbase(&self) -> Result<(), Error> {
+	/// Validate the coinbase outputs generated by miners. Entails 2 main checks:
+	///
+	/// * That the sum of all coinbase-marked outputs equal the supply.
+	/// * That the sum of blinding factors for all coinbase-marked outputs match
+	///   the coinbase-marked kernels.
+	pub fn verify_coinbase(&self) -> Result<(), Error> {
 		let cb_outs = self.outputs
 			.iter()
 			.filter(|out| out.features.contains(OutputFeatures::COINBASE_OUTPUT))
@@ -804,382 +802,5 @@ impl Block {
 			return Err(Error::CoinbaseSumMismatch);
 		}
 		Ok(())
-	}
-}
-
-#[cfg(test)]
-mod test {
-	use super::*;
-	use core::Transaction;
-	use core::build::{self, input, output, with_fee};
-	use core::test::{tx1i2o, tx2i1o};
-	use keychain::{Identifier, Keychain};
-	use consensus::{BLOCK_OUTPUT_WEIGHT, MAX_BLOCK_WEIGHT};
-	use std::time::Instant;
-
-	use util::secp;
-
-	// utility to create a block without worrying about the key or previous
-	// header
-	fn new_block(
-		txs: Vec<&Transaction>,
-		keychain: &Keychain,
-		previous_header: &BlockHeader,
-	) -> Block {
-		let key_id = keychain.derive_key_id(1).unwrap();
-		Block::new(&previous_header, txs, keychain, &key_id, Difficulty::one()).unwrap()
-	}
-
-	// utility producing a transaction that spends an output with the provided
-	// value and blinding key
-	fn txspend1i1o(
-		v: u64,
-		keychain: &Keychain,
-		key_id1: Identifier,
-		key_id2: Identifier,
-	) -> Transaction {
-		build::transaction(
-			vec![input(v, key_id1), output(3, key_id2), with_fee(2)],
-			&keychain,
-		).unwrap()
-	}
-
-	// Too slow for now #[test]
-	// TODO: make this fast enough or add similar but faster test?
-	#[allow(dead_code)]
-	fn too_large_block() {
-		let keychain = Keychain::from_random_seed().unwrap();
-		let max_out = MAX_BLOCK_WEIGHT / BLOCK_OUTPUT_WEIGHT;
-
-		let zero_commit = secp_static::commit_to_zero_value();
-
-		let mut pks = vec![];
-		for n in 0..(max_out + 1) {
-			pks.push(keychain.derive_key_id(n as u32).unwrap());
-		}
-
-		let mut parts = vec![];
-		for _ in 0..max_out {
-			parts.push(output(5, pks.pop().unwrap()));
-		}
-
-		let now = Instant::now();
-		parts.append(&mut vec![input(500000, pks.pop().unwrap()), with_fee(2)]);
-		let mut tx = build::transaction(parts, &keychain).unwrap();
-		println!("Build tx: {}", now.elapsed().as_secs());
-
-		let prev = BlockHeader::default();
-		let b = new_block(vec![&mut tx], &keychain, &prev);
-		assert!(b.validate(&zero_commit, &zero_commit).is_err());
-	}
-
-	#[test]
-	// block with no inputs/outputs/kernels
-	// no fees, no reward, no coinbase
-	fn very_empty_block() {
-		let b = Block {
-			header: BlockHeader::default(),
-			inputs: vec![],
-			outputs: vec![],
-			kernels: vec![],
-		};
-
-		assert_eq!(
-			b.verify_coinbase(),
-			Err(Error::Secp(secp::Error::IncorrectCommitSum))
-		);
-	}
-
-	#[test]
-	// builds a block with a tx spending another and check that cut_through occurred
-	fn block_with_cut_through() {
-		let keychain = Keychain::from_random_seed().unwrap();
-		let key_id1 = keychain.derive_key_id(1).unwrap();
-		let key_id2 = keychain.derive_key_id(2).unwrap();
-		let key_id3 = keychain.derive_key_id(3).unwrap();
-
-		let zero_commit = secp_static::commit_to_zero_value();
-
-		let mut btx1 = tx2i1o();
-		let mut btx2 = build::transaction(
-			vec![input(7, key_id1), output(5, key_id2.clone()), with_fee(2)],
-			&keychain,
-		).unwrap();
-
-		// spending tx2 - reuse key_id2
-
-		let mut btx3 = txspend1i1o(5, &keychain, key_id2.clone(), key_id3);
-		let prev = BlockHeader::default();
-		let b = new_block(vec![&mut btx1, &mut btx2, &mut btx3], &keychain, &prev);
-
-		// block should have been automatically compacted (including reward
-		// output) and should still be valid
-		b.validate(&zero_commit, &zero_commit).unwrap();
-		assert_eq!(b.inputs.len(), 3);
-		assert_eq!(b.outputs.len(), 3);
-	}
-
-	#[test]
-	fn empty_block_with_coinbase_is_valid() {
-		let keychain = Keychain::from_random_seed().unwrap();
-		let zero_commit = secp_static::commit_to_zero_value();
-		let prev = BlockHeader::default();
-		let b = new_block(vec![], &keychain, &prev);
-
-		assert_eq!(b.inputs.len(), 0);
-		assert_eq!(b.outputs.len(), 1);
-		assert_eq!(b.kernels.len(), 1);
-
-		let coinbase_outputs = b.outputs
-			.iter()
-			.filter(|out| out.features.contains(OutputFeatures::COINBASE_OUTPUT))
-			.map(|o| o.clone())
-			.collect::<Vec<_>>();
-		assert_eq!(coinbase_outputs.len(), 1);
-
-		let coinbase_kernels = b.kernels
-			.iter()
-			.filter(|out| out.features.contains(KernelFeatures::COINBASE_KERNEL))
-			.map(|o| o.clone())
-			.collect::<Vec<_>>();
-		assert_eq!(coinbase_kernels.len(), 1);
-
-		// the block should be valid here (single coinbase output with corresponding
-		// txn kernel)
-		assert!(b.validate(&zero_commit, &zero_commit).is_ok());
-	}
-
-	#[test]
-	// test that flipping the COINBASE_OUTPUT flag on the output features
-	// invalidates the block and specifically it causes verify_coinbase to fail
-	// additionally verifying the merkle_inputs_outputs also fails
-	fn remove_coinbase_output_flag() {
-		let keychain = Keychain::from_random_seed().unwrap();
-		let zero_commit = secp_static::commit_to_zero_value();
-		let prev = BlockHeader::default();
-		let mut b = new_block(vec![], &keychain, &prev);
-
-		assert!(
-			b.outputs[0]
-				.features
-				.contains(OutputFeatures::COINBASE_OUTPUT)
-		);
-		b.outputs[0]
-			.features
-			.remove(OutputFeatures::COINBASE_OUTPUT);
-
-		assert_eq!(b.verify_coinbase(), Err(Error::CoinbaseSumMismatch));
-		assert!(b.verify_sums(&zero_commit, &zero_commit).is_ok());
-		assert_eq!(
-			b.validate(&zero_commit, &zero_commit),
-			Err(Error::CoinbaseSumMismatch)
-		);
-	}
-
-	#[test]
-	// test that flipping the COINBASE_KERNEL flag on the kernel features
-	// invalidates the block and specifically it causes verify_coinbase to fail
-	fn remove_coinbase_kernel_flag() {
-		let keychain = Keychain::from_random_seed().unwrap();
-		let zero_commit = secp_static::commit_to_zero_value();
-		let prev = BlockHeader::default();
-		let mut b = new_block(vec![], &keychain, &prev);
-
-		assert!(
-			b.kernels[0]
-				.features
-				.contains(KernelFeatures::COINBASE_KERNEL)
-		);
-		b.kernels[0]
-			.features
-			.remove(KernelFeatures::COINBASE_KERNEL);
-
-		assert_eq!(
-			b.verify_coinbase(),
-			Err(Error::Secp(secp::Error::IncorrectCommitSum))
-		);
-
-		assert_eq!(
-			b.validate(&zero_commit, &zero_commit),
-			Err(Error::Secp(secp::Error::IncorrectCommitSum))
-		);
-	}
-
-	#[test]
-	fn serialize_deserialize_block() {
-		let keychain = Keychain::from_random_seed().unwrap();
-		let prev = BlockHeader::default();
-		let b = new_block(vec![], &keychain, &prev);
-
-		let mut vec = Vec::new();
-		ser::serialize(&mut vec, &b).expect("serialization failed");
-		let b2: Block = ser::deserialize(&mut &vec[..]).unwrap();
-
-		assert_eq!(b.header, b2.header);
-		assert_eq!(b.inputs, b2.inputs);
-		assert_eq!(b.outputs, b2.outputs);
-		assert_eq!(b.kernels, b2.kernels);
-	}
-
-	#[test]
-	fn empty_block_serialized_size() {
-		let keychain = Keychain::from_random_seed().unwrap();
-		let prev = BlockHeader::default();
-		let b = new_block(vec![], &keychain, &prev);
-		let mut vec = Vec::new();
-		ser::serialize(&mut vec, &b).expect("serialization failed");
-		let target_len = 1_216;
-		assert_eq!(vec.len(), target_len,);
-	}
-
-	#[test]
-	fn block_single_tx_serialized_size() {
-		let keychain = Keychain::from_random_seed().unwrap();
-		let tx1 = tx1i2o();
-		let prev = BlockHeader::default();
-		let b = new_block(vec![&tx1], &keychain, &prev);
-		let mut vec = Vec::new();
-		ser::serialize(&mut vec, &b).expect("serialization failed");
-		let target_len = 2_796;
-		assert_eq!(vec.len(), target_len);
-	}
-
-	#[test]
-	fn empty_compact_block_serialized_size() {
-		let keychain = Keychain::from_random_seed().unwrap();
-		let prev = BlockHeader::default();
-		let b = new_block(vec![], &keychain, &prev);
-		let mut vec = Vec::new();
-		ser::serialize(&mut vec, &b.as_compact_block()).expect("serialization failed");
-		let target_len = 1_224;
-		assert_eq!(vec.len(), target_len,);
-	}
-
-	#[test]
-	fn compact_block_single_tx_serialized_size() {
-		let keychain = Keychain::from_random_seed().unwrap();
-		let tx1 = tx1i2o();
-		let prev = BlockHeader::default();
-		let b = new_block(vec![&tx1], &keychain, &prev);
-		let mut vec = Vec::new();
-		ser::serialize(&mut vec, &b.as_compact_block()).expect("serialization failed");
-		let target_len = 1_230;
-		assert_eq!(vec.len(), target_len,);
-	}
-
-	#[test]
-	fn block_10_tx_serialized_size() {
-		let keychain = Keychain::from_random_seed().unwrap();
-		global::set_mining_mode(global::ChainTypes::Mainnet);
-
-		let mut txs = vec![];
-		for _ in 0..10 {
-			let tx = tx1i2o();
-			txs.push(tx);
-		}
-		let prev = BlockHeader::default();
-		let b = new_block(txs.iter().collect(), &keychain, &prev);
-		let mut vec = Vec::new();
-		ser::serialize(&mut vec, &b).expect("serialization failed");
-		let target_len = 17_016;
-		assert_eq!(vec.len(), target_len,);
-	}
-
-	#[test]
-	fn compact_block_10_tx_serialized_size() {
-		let keychain = Keychain::from_random_seed().unwrap();
-
-		let mut txs = vec![];
-		for _ in 0..10 {
-			let tx = tx1i2o();
-			txs.push(tx);
-		}
-		let prev = BlockHeader::default();
-		let b = new_block(txs.iter().collect(), &keychain, &prev);
-		let mut vec = Vec::new();
-		ser::serialize(&mut vec, &b.as_compact_block()).expect("serialization failed");
-		let target_len = 1_284;
-		assert_eq!(vec.len(), target_len,);
-	}
-
-	#[test]
-	fn compact_block_hash_with_nonce() {
-		let keychain = Keychain::from_random_seed().unwrap();
-		let tx = tx1i2o();
-		let prev = BlockHeader::default();
-		let b = new_block(vec![&tx], &keychain, &prev);
-		let cb1 = b.as_compact_block();
-		let cb2 = b.as_compact_block();
-
-		// random nonce will not affect the hash of the compact block itself
-		// hash is based on header POW only
-		assert!(cb1.nonce != cb2.nonce);
-		assert_eq!(b.hash(), cb1.hash());
-		assert_eq!(cb1.hash(), cb2.hash());
-
-		assert!(cb1.kern_ids[0] != cb2.kern_ids[0]);
-
-		// check we can identify the specified kernel from the short_id
-		// correctly in both of the compact_blocks
-		assert_eq!(
-			cb1.kern_ids[0],
-			tx.kernels[0].short_id(&cb1.hash(), cb1.nonce)
-		);
-		assert_eq!(
-			cb2.kern_ids[0],
-			tx.kernels[0].short_id(&cb2.hash(), cb2.nonce)
-		);
-	}
-
-	#[test]
-	fn convert_block_to_compact_block() {
-		let keychain = Keychain::from_random_seed().unwrap();
-		let tx1 = tx1i2o();
-		let prev = BlockHeader::default();
-		let b = new_block(vec![&tx1], &keychain, &prev);
-		let cb = b.as_compact_block();
-
-		assert_eq!(cb.out_full.len(), 1);
-		assert_eq!(cb.kern_full.len(), 1);
-		assert_eq!(cb.kern_ids.len(), 1);
-
-		assert_eq!(
-			cb.kern_ids[0],
-			b.kernels
-				.iter()
-				.find(|x| !x.features.contains(KernelFeatures::COINBASE_KERNEL))
-				.unwrap()
-				.short_id(&cb.hash(), cb.nonce)
-		);
-	}
-
-	#[test]
-	fn hydrate_empty_compact_block() {
-		let keychain = Keychain::from_random_seed().unwrap();
-		let prev = BlockHeader::default();
-		let b = new_block(vec![], &keychain, &prev);
-		let cb = b.as_compact_block();
-		let hb = Block::hydrate_from(cb, vec![]);
-		assert_eq!(hb.header, b.header);
-		assert_eq!(hb.outputs, b.outputs);
-		assert_eq!(hb.kernels, b.kernels);
-	}
-
-	#[test]
-	fn serialize_deserialize_compact_block() {
-		let b = CompactBlock {
-			header: BlockHeader::default(),
-			nonce: 0,
-			out_full: vec![],
-			kern_full: vec![],
-			kern_ids: vec![ShortId::zero()],
-		};
-
-		let mut vec = Vec::new();
-		ser::serialize(&mut vec, &b).expect("serialization failed");
-		let b2: CompactBlock = ser::deserialize(&mut &vec[..]).unwrap();
-
-		assert_eq!(b.header, b2.header);
-		assert_eq!(b.kern_ids, b2.kern_ids);
 	}
 }
