+++ pool/Cargo.toml
@@ -16,3 +16,6 @@ time = "0.1"
 grin_core = { path = "../core" }
 grin_keychain = { path = "../keychain" }
 grin_util = { path = "../util" }
+
+[dev-dependencies]
+grin_wallet = { path = "../wallet" }
